+++ teneto/utils/__init__.py
@@ -2,4 +2,5 @@
 Import utility funcitons
 """
 
+from .bidsutils import * 
 from .utils import *